테이블 스키마 정리 및 버그 개선

+++ 2.Modules/openapi/Entity/ApiParameter.cs
@@ -14,7 +14,7 @@ namespace openapi.Entity
         public string DefaultValue { get; set; } = string.Empty;
 
         [JsonProperty("Length")]
-        public int Length { get; set; } = 0;
+        public long Length { get; set; } = 0;
 
         [JsonProperty("RequiredYN")]
         public bool RequiredYN { get; set; } = false;
